Move `ColonyShip` loading code to `SavedGameFile`
